kubelet: Move container reference manager to pkg/kubelet/container.
This enables other package to use it, such as docktools.
